extract filterPanel
