+++ client/src/components/deposits/ProductForm.tsx
@@ -18,7 +18,8 @@ import {
 } from '@/api/generated/zod.gen'
 import { useAppForm } from '@/components/form'
 import { Alert, AlertDescription, AlertTitle } from '@/components/ui/alert'
-import { ProductType, formatCoordinates } from '@/lib/utils'
+import { ProductType } from '@/lib/utils'
+import { formatCoordinates } from '@/utils/formatter'
 import { useStore } from '@tanstack/react-form'
 import { useMutation } from '@tanstack/react-query'
 import { useNavigate } from '@tanstack/react-router'
